wildfire-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-cloud-bigquery==3.25.0
db-dtypes==1.2.0
requests==2.32.3